(doc) Change commit message instructions to '(doc)' to match puppet.git
The CONTRIBUTING.md file in the main puppet repo suggests
that minor documentation edits can have commit messages prefixed with '(doc)'
in lieu of filing a JIRA ticket.  This file said '(docs)'.
I checked the git history in the main repo and found a lot more commits
with '(doc)' than '(docs)', so change this one for consistency.